Rename main zshrc to zshrc.ohmyzsh
This is to allow for a general zsh config that doesn't change between
machines. Instead all local changes should go into the ~/.zshrc file,
which in turn will source the standard ~/.zshrc.ohmyzsh file.

The reason .zshrc is kept as the local file is that it's assumed in a
lot of pre-existing scripts.